refactor: :recycle: moved SleepyBaby as package
